Avoid negative values other than -1 in ssize_t / change naming scheme for internal methods
ssize_t technically must only be able to store values from -1 to
SIZE_MAX.

To simplify the building of frames in the process, this commit lets
safer_snprintf (renamed to _rtm_snprintf) and the json encode calls
return pointers to the end of the string instead of the length of the
bytes written. The functions also silently return NULL if they are
passed a NULL pointer. This allows to chain the calls and only check
once after building a frame whether this succeeded or not.

Furthermore, this commit unifies the naming convention for internal
methods: They all have the _rtm-prefix now, making it easy to
distinguish them from other libraries and stdlib calls. It also adds
documentation to these methods.

Finally, the commit fixes some bugs that were too easily fixed to
justify a separate commit.

+++ core/src/rtm_apple_ssl.c
@@ -20,6 +20,8 @@ static OSStatus read_from_socket(SSLConnectionRef connection, void *data, size_t
   if (status > 0) {
     *len = (size_t) status;
     if (requested_sz > *len)
+      // sic! Apple SSL requires to return an error if fewer bytes have been
+      // read than requested.
       return errSSLWouldBlock;
     else
       return noErr;
@@ -51,10 +53,12 @@ static OSStatus write_to_socket(SSLConnectionRef connection, const void *data, s
   ASSERT_NOT_NULL(len);
 
   size_t to_write_sz = *len;
-  ssize_t status = write(fd, data, to_write_sz);
+  int status = write(fd, data, to_write_sz);
 
   if (status > 0) {
     *len = (size_t) status;
+    // sic! Apple SSL requires to return an error if fewer bytes have been
+    // written than requested.
     if (to_write_sz > *len)
       return errSSLWouldBlock;
     else
